Create pythonprojectfinal.py
Protect Overview:
This project is a poetry generator, where different words that could make up poetic phrases are generators for the user to arrange on their own. The small piece of the project that I managed to perform was the actual generation of different words in sentences. These words range from different types (like nouns and verbs) to different tones (like bright and dark). These randomly generated words from a given list can then be combined to perhaps create a poetic phrase by the user.

Project Narrative:
For the previous milestones, I essentially took a basic outline for milestone one and then built upon that for milestone two. They both contained the same overall concept in terms of code with the word generation, but they contained different lists and segments. Throughout the lectures, the ones that were helpful were the text generator lecture, as well as the objects and classes, since they both enabled me to in a way understand what I was actually attempting to convey. However, a couple confused me and it seemed like the progression was too quick to grasp each new tactic through coding and then python. I think what surprised me the most throughout the lectures and work is that there is a specific way to do things, and once the ways are learned, it is easy to continue building and growing in terms of coding. If I was to approach the project again, it would ideally be in a setting where I can actually interact and observe hands-on various other fundamentals so that I could expand the project more. 

Corrections:
The main comment I responded to was that the milestones were a good start, which I responded to by elaborating more and adding more exact code. The other part of the comment that I responded to was that part of the code with "**" was incorrect, so I responded by removing that segment.